move WebTestCase to "Functional" namespace instead of "Units"
